Improve Docker load speed
To improve the speed of docker push, we changed the order in which the
layers are processed. Speed is gain when we do not have to push many
layers to Docker, thus it makes sense to start checking the layer in
reverse order to minimize calls to Docker when only the lop layer has
changed for instance.